% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PERAssessmentGuides-Package.R
\docType{package}
\name{PERAssessmentGuides-package}
\alias{PERAssessmentGuides-package}
\title{PER Assessment Guide Templates:}
\description{
This package contains an RMarkdown template/skeleton for generating assessment guides

\tabular{ll}{
  Package: \tab granovaGG\cr
  Version: \tab 0.1\cr
  License: \tab GPL (>= 2)\cr
}
}
\author{
Brian A. Danielak \email{brian@briandk.com}\cr
}

